Lots of optimizations for full-speed mode, make button bar resize to fit, and various attempts to make shell layout more sane (saving at checkpoint)

+++ src/v9t9/emulator/hardware/memory/mmio/GplMmio.java
@@ -40,7 +40,7 @@ public class GplMmio implements ConsoleMmioReader, ConsoleMmioWriter {
         return (short) (addr+1 & 0x1fff | gromaddr & 0xe000);
     }
     
-    public short getAddr() {
+    public int getAddr() {
         return gromaddr;
     }
